# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AreaStatus": {
        "additionalProperties": false,
        "properties": {
          "area": {
            "title": "Area",
            "type": "string"
          },
          "credits": {
            "title": "Credits",
            "type": "number"
          },
          "satisfied": {
            "title": "Satisfied",
            "type": "boolean"
          }
        },
        "required": [
          "area",
          "credits",
          "satisfied"
        ],
        "title": "AreaStatus",
        "type": "object"
      },
...
      "CourseListRead": {
        "additionalProperties": false,
        "properties": {
          "courses": {
            "items": {
              "$ref": "#/components/schemas/CourseSummaryRead"
            },
            "title": "Courses",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "courses",
          "page",
          "size",
          "total"
        ],
        "title": "CourseListRead",
        "type": "object"
      },
      "CourseSummaryRead": {
        "additionalProperties": false,
        "properties": {
          "averageRating": {
            "title": "Averagerating",
            "type": "number"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "courseCode": {
            "title": "Coursecode",
            "type": "string"
          },
          "credits": {
            "title": "Credits",
            "type": "number"
          },
          "grade": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "popularityScore": {
            "title": "Popularityscore",
            "type": "number"
          },
          "professors": {
            "items": {
              "type": "string"
            },
            "title": "Professors",
            "type": "array"
          },
          "reviewCount": {
            "title": "Reviewcount",
            "type": "integer"
          },
          "sectionCount": {
            "title": "Sectioncount",
            "type": "integer"
          }
        },
        "required": [
          "courseCode",
          "name",
          "category",
          "credits",
          "grade",
          "sectionCount",
          "professors",
          "averageRating",
          "reviewCount",
          "popularityScore"
        ],
        "title": "CourseSummaryRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RequirementEvaluation": {
        "additionalProperties": false,
        "properties": {
          "areaStatus": {
            "items": {
              "$ref": "#/components/schemas/AreaStatus"
            },
            "title": "Areastatus",
            "type": "array"
          },
          "creditStatus": {
            "items": {
              "$ref": "#/components/schemas/RequirementStatus"
            },
            "title": "Creditstatus",
            "type": "array"
          },
          "manualReviewItems": {
            "items": {
              "type": "string"
            },
            "title": "Manualreviewitems",
            "type": "array"
          },
          "missingRequirements": {
            "items": {
              "type": "string"
            },
            "title": "Missingrequirements",
            "type": "array"
          },
          "requiredCourseStatus": {
            "items": {
              "$ref": "#/components/schemas/RequirementStatus"
            },
            "title": "Requiredcoursestatus",
            "type": "array"
          }
        },
        "required": [
          "creditStatus",
          "areaStatus",
          "requiredCourseStatus",
          "missingRequirements",
          "manualReviewItems"
        ],
        "title": "RequirementEvaluation",
        "type": "object"
      },
      "RequirementProfile": {
        "additionalProperties": false,
        "properties": {
          "admissionYear": {
            "maximum": 2100.0,
            "minimum": 1990.0,
            "title": "Admissionyear",
            "type": "integer"
          },
          "departmentId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Departmentid",
            "type": "string"
          },
          "programPath": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Programpath",
            "type": "string"
          },
          "studentType": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Studenttype",
            "type": "string"
          }
        },
        "required": [
          "admissionYear",
          "departmentId",
          "studentType",
          "programPath"
        ],
        "title": "RequirementProfile",
        "type": "object"
      },
      "RequirementRecommendation": {
        "additionalProperties": false,
        "properties": {
          "missingRequirements": {
            "items": {
              "type": "string"
            },
            "title": "Missingrequirements",
            "type": "array"
          },
          "recommendedCourses": {
            "items": {
              "$ref": "#/components/schemas/CourseSummaryRead"
            },
            "title": "Recommendedcourses",
            "type": "array"
          }
        },
        "required": [
          "missingRequirements",
          "recommendedCourses"
        ],
        "title": "RequirementRecommendation",
        "type": "object"
      },
      "RequirementRuleList": {
        "additionalProperties": false,
        "properties": {
          "asOf": {
            "title": "Asof",
            "type": "string"
          },
          "manualReviewItems": {
            "items": {
              "type": "string"
            },
            "title": "Manualreviewitems",
            "type": "array"
          },
          "rules": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rules",
            "type": "array"
          }
        },
        "required": [
          "rules",
          "manualReviewItems",
          "asOf"
        ],
        "title": "RequirementRuleList",
        "type": "object"
      },
      "RequirementSource": {
        "additionalProperties": false,
        "properties": {
          "effectiveDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effectivedate"
          },
          "sourceId": {
            "title": "Sourceid",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "verifiedAt": {
            "title": "Verifiedat",
            "type": "string"
          }
        },
        "required": [
          "sourceId",
          "title",
          "url",
          "verifiedAt"
        ],
        "title": "RequirementSource",
        "type": "object"
      },
      "RequirementStatus": {
        "additionalProperties": false,
        "properties": {
          "current": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "missing": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Missing",
            "type": "array"
          },
          "required": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required"
          },
          "satisfied": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Satisfied"
          }
        },
        "required": [
          "kind",
          "required",
          "current",
          "satisfied"
        ],
        "title": "RequirementStatus",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "PL-timeTabler API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/courses": {
      "get": {
        "operationId": "list_courses_api_v1_courses_get",
        "parameters": [
          {
            "in": "query",
            "name": "semester",
            "required": false,
            "schema": {
              "default": "2026-1",
              "title": "Semester",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "in": "query",
            "name": "professor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Professor"
            }
          },
          {
            "in": "query",
            "name": "departmentId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Departmentid"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "area",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Area"
            }
          },
          {
            "in": "query",
            "name": "grade",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 4,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Grade"
            }
          },
          {
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[월화수목금토일]$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "NAME",
              "pattern": "^(NAME|POPULARITY|RATING|REVIEWS)$",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "ASC",
              "pattern": "^(ASC|DESC)$",
              "title": "Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 2000,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseListRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Courses",
        "tags": [
          "courses"
        ]
      }
    },
...
    "/api/v1/requirements/evaluate": {
      "post": {
        "operationId": "evaluate_requirements_api_v1_requirements_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementProfile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementEvaluation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Requirements",
        "tags": [
          "requirements"
        ]
      }
    },
    "/api/v1/requirements/recommendations": {
      "get": {
        "operationId": "requirement_recommendations_api_v1_requirements_recommendations_get",
        "parameters": [
          {
            "in": "query",
            "name": "semester",
            "required": false,
            "schema": {
              "default": "2026-1",
              "title": "Semester",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "admissionYear",
            "required": true,
            "schema": {
              "maximum": 2100,
              "minimum": 1990,
              "title": "Admissionyear",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "departmentId",
            "required": true,
            "schema": {
              "title": "Departmentid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "studentType",
            "required": true,
            "schema": {
              "title": "Studenttype",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "programPath",
            "required": true,
            "schema": {
              "title": "Programpath",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementRecommendation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Requirement Recommendations",
        "tags": [
          "requirements"
        ]
      }
    },
    "/api/v1/requirements/rules": {
      "get": {
        "operationId": "list_requirement_rules_api_v1_requirements_rules_get",
        "parameters": [
          {
            "in": "query",
            "name": "admissionYear",
            "required": true,
            "schema": {
              "maximum": 2100,
              "minimum": 1990,
              "title": "Admissionyear",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "departmentId",
            "required": true,
            "schema": {
              "title": "Departmentid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "studentType",
            "required": true,
            "schema": {
              "title": "Studenttype",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "programPath",
            "required": true,
            "schema": {
              "title": "Programpath",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementRuleList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Requirement Rules",
        "tags": [
          "requirements"
        ]
      }
    },
    "/api/v1/requirements/sources/{source_id}": {
      "get": {
        "operationId": "get_requirement_source_api_v1_requirements_sources__source_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementSource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Requirement Source",
        "tags": [
          "requirements"
        ]
      }
    },
...
  }
}